requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.126.0
pypdf
langchain-text-splitters
sentence-transformers
faiss-cpu
numpy
groq
requests
gdown==5.1.0
